--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/IdentityProviderApplication.java
@@ -1,13 +1,0 @@
-package com.team.authorizeservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class IdentityProviderApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(IdentityProviderApplication.class, args);
-	}
-}
