--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/persistence/model/User.java
@@ -1,13 +1,0 @@
-package com.team.authorizeservice.persistence.model;
-
-import lombok.Value;
-
-@Value
-public class User {
-  Long id;
-  String name;
-  String email;
-  String password;
-  boolean active;
-  String role;
-}
