--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/rest/dto/AuthenticationDto.java
@@ -1,9 +1,0 @@
-package com.team.authorizeservice.rest.dto;
-
-import lombok.Value;
-
-@Value
-public class AuthenticationDto {
-  String email;
-  String password;
-}
