--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/rest/dto/RequestMetadata.java
@@ -1,17 +1,0 @@
-package com.team.authorizeservice.rest.dto;
-
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.Value;
-
-@Value
-public class RequestMetadata {
-  String remoteAddress;
-  String userAgent;
-
-  public RequestMetadata(HttpServletRequest request) {
-    remoteAddress = request.getRemoteAddr();
-    userAgent = request.getHeader("user-agent");
-  }
-}
-
-
