--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/security/jwt/exception/JwtAuthenticationException.java
@@ -1,13 +1,0 @@
-package com.team.authorizeservice.security.jwt.exception;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class JwtAuthenticationException extends AuthenticationException {
-  public JwtAuthenticationException(String msg, Throwable cause) {
-    super(msg, cause);
-  }
-
-  public JwtAuthenticationException(String msg) {
-    super(msg);
-  }
-}
