--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/security/jwt/exception/TokenResolvingException.java
@@ -1,11 +1,0 @@
-package com.team.authorizeservice.security.jwt.exception;
-
-public class TokenResolvingException extends RuntimeException {
-  public TokenResolvingException() {
-    super();
-  }
-
-  public TokenResolvingException(String message) {
-    super(message);
-  }
-}
