--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/service/api/AuthenticationService.java
@@ -1,8 +1,0 @@
-package com.team.authorizeservice.service.api;
-
-import com.team.authorizeservice.security.details.ProjectionUserDetails;
-import org.springframework.security.core.Authentication;
-
-public interface AuthenticationService {
-  ProjectionUserDetails authenticate(String email, String password);
-}
