--- backend/store/identity-provider/src/main/java/com/team/authorizeservice/service/api/RefreshSessionService.java
@@ -1,9 +1,0 @@
-package com.team.authorizeservice.service.api;
-
-import com.team.authorizeservice.persistence.model.RefreshSession;
-
-public interface RefreshSessionService {
-  void save(RefreshSession refreshSession);
-  void deleteByToken(String refreshToken);
-  boolean existsByToken(String refreshToken);
-}
