+++ backend/store/identity-provider/src/main/java/com/team/identityprovider/IdentityProviderApplication.java
@@ -1,0 +1,13 @@
+package com.team.identityprovider;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class IdentityProviderApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(IdentityProviderApplication.class, args);
+	}
+}
