+++ backend/store/identity-provider/src/main/java/com/team/identityprovider/persistence/model/User.java
@@ -1,0 +1,13 @@
+package com.team.identityprovider.persistence.model;
+
+import lombok.Value;
+
+@Value
+public class User {
+  Long id;
+  String name;
+  String email;
+  String password;
+  boolean active;
+  String role;
+}
