+++ backend/store/identity-provider/src/main/java/com/team/identityprovider/rest/dto/AuthenticationDto.java
@@ -1,0 +1,9 @@
+package com.team.identityprovider.rest.dto;
+
+import lombok.Value;
+
+@Value
+public class AuthenticationDto {
+  String email;
+  String password;
+}
