+++ backend/store/identity-provider/src/main/java/com/team/identityprovider/security/details/ProjectionUserDetails.java
@@ -1,0 +1,71 @@
+package com.team.identityprovider.security.details;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.AllArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+@AllArgsConstructor
+public class ProjectionUserDetails implements UserDetails {
+  private final Long id;
+  private final String name;
+  private final String email;
+  private final String password;
+  private final boolean active;
+  private final Collection<? extends GrantedAuthority> authorities;
+
+  @JsonIgnore
+  public Long getId() {
+    return id;
+  }
+
+  @JsonIgnore
+  @Override
+  public String getPassword() {
+    return password;
+  }
+
+  @JsonIgnore
+  @Override
+  public String getUsername() {
+    return email;
+  }
+
+  @Override
+  public boolean isEnabled() {
+    return active;
+  }
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    return authorities;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public String getEmail() {
+    return email;
+  }
+
+  @JsonIgnore
+  @Override
+  public boolean isAccountNonExpired() {
+    return true;
+  }
+
+  @JsonIgnore
+  @Override
+  public boolean isAccountNonLocked() {
+    return true;
+  }
+
+  @JsonIgnore
+  @Override
+  public boolean isCredentialsNonExpired() {
+    return true;
+  }
+}
