+++ backend/store/identity-provider/src/main/java/com/team/identityprovider/security/jwt/exception/JwtAuthenticationException.java
@@ -1,0 +1,13 @@
+package com.team.identityprovider.security.jwt.exception;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class JwtAuthenticationException extends AuthenticationException {
+  public JwtAuthenticationException(String msg, Throwable cause) {
+    super(msg, cause);
+  }
+
+  public JwtAuthenticationException(String msg) {
+    super(msg);
+  }
+}
