+++ backend/store/identity-provider/src/main/java/com/team/identityprovider/security/jwt/exception/TokenResolvingException.java
@@ -1,0 +1,11 @@
+package com.team.identityprovider.security.jwt.exception;
+
+public class TokenResolvingException extends RuntimeException {
+  public TokenResolvingException() {
+    super();
+  }
+
+  public TokenResolvingException(String message) {
+    super(message);
+  }
+}
