+++ backend/store/identity-provider/src/main/java/com/team/identityprovider/service/api/AuthenticationService.java
@@ -1,0 +1,7 @@
+package com.team.identityprovider.service.api;
+
+import com.team.identityprovider.security.details.ProjectionUserDetails;
+
+public interface AuthenticationService {
+  ProjectionUserDetails authenticate(String email, String password);
+}
